requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
greenlet==3.1.1
mysql-connector-python==9.2.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
typing_extensions==4.12.2
